requirements.txt
einops
matplotlib
pillow